Removed public subnet and ip requirements
Removed the requirement for the DataSync in-cloud agent to be launched in a public subnet and to have a public ipv4 address. The source subnet can now be a private subnet, with a route to the Internet via a NAT gateway, and a public ipv4 address is no longer automatically assigned to the DataSync in-cloud agent.